Large refactor to remove `cirq` from dependencies (#334)
To help anyone wanting to review that in the future 
below is an explanation of the important changes.

-------

## `ScheduledCircuit`

The `ScheduledCircuit` class has been reworked extensively. It now
stores a list of `Moment` instances (a new class introduced by this PR
and that nearly mimic `cirq.Moment` semantic, but using `stim.Circuit`),
a `Schedule` instance (a new class introduced by this PR to avoid having
`list[int]` moving around and re-doing checks on them) and a global
qubit map.

A few notes on that:

`ScheduledCircuit` does not store a full circuit anymore. That means
that the circuit needs to be built each time it is needed (see
`ScheduledCircuit.get_circuit` and
`ScheduledCircuit.get_repeated_circuit`).

`stim.Circuit` handles qubits differently than `cirq.Circuit` and that
shows in the new interface. Basically, each operation `cirq.Circuit`
contained directly the qubits (`cirq.GridQubit` instances) whereas
`stim.Circuit` operations are applied on "targets" that, for quantum
gates, are integer indices that index a qubit map defined using
`QUBIT_COORDS` instructions. That means that there are now two different
(in terms of behaviour **and** runtime cost) operations to remap qubits:

1. Changing the qubits the circuit is applied on, which can be done
efficiently simply by changing the qubit map and in particular without
changing any instruction.
2. Changing the qubit indices, that require to change both the qubit map
**and** each instruction, which basically require to create a new
circuit by iterating on all the instructions.

Operation 1. is the easiest to understand. Operation 2. is needed when
several `ScheduledCircuit` have been created independently of each other
and will eventually be merged together (which is a huge portion of what
we are doing).

## Removing `cirq -> stim` code and updating existing code

All the code related to translating from `cirq` to `stim` has been
removed.

All the code that was using `ScheduledCircuit` has been updated to use
the "new" interface (which is mostly a change on how to create a
`ScheduleCircuit`, how to manage qubits when merging and return types).
That includes `tqec.plaquette`, `tqec.compile` and parts of
`tqec.circuit`.

## Benchmarks

Before the change, we had the following performance (from #329):

![image](https://github.com/user-attachments/assets/c4c44259-de7b-4997-b66b-952eb85d102c)

Now, **without the measurement map for the moment**, we have:

![image](https://github.com/user-attachments/assets/4b28c6ae-a0b2-4a89-a53f-40849c9a0e6f)

A few remarks:

- detector computation is still the same cost, which is expected and a
nice indication that we might still have a correct circuit.
- circuit construction was nearly 25 seconds, it is now 3 seconds
(without `MeasurementMap`, but `MeasurementMap` took at most 6.5 seconds
in the first implementation, so we still have a nice improvement).

---------

Co-authored-by: Yiming Zhang <61700160+inmzhang@users.noreply.github.com>